Small refactor
Separated base logic from panel to reutilize it in another panel
