+++ LeapMotionBlender/UI/Panels/__init__.py
@@ -1,0 +1,1 @@
+from .hand_select import HandSelect
