+++ LeapMotionBlender/__init__.py
@@ -31,10 +31,10 @@ from .general_helpers import register_with_extras, unregister_with_extras
 from . import communicator
 from .Operators import ForceStart
 from .settingsPanel import SettingsPanel
-from .UI import BoneSelectPanel
+from .UI import HandSelect
 
 
-classes = (SettingsPanel, ForceStart, BoneSelectPanel)
+classes = (SettingsPanel, ForceStart, HandSelect)
 
 def register():
